Use cookies directly from the webview
Originally, cookies from the webview were always copied to the
WebRequest. When getting the cookies from the webiew only those
for the current URL can be retrieved. This means that cookies of other
domains were never copied to the WebRequest for upload and download.

This commit adds the option to use a HTTP request constructed by the
webview, which automatically adds any available cookies for that
request.

Fixed by using the BrowserHttp implementation of IWebRequestCreate
which shares the cookie store with the webview. See also
http://stackoverflow.com/questions/4212713/grabbing-cookies-in-web-browser-control-wp7

This functionality is enabled by a new useBrowserHttp flag.  We need
this to be optional because:

- It removes the possibility of adding cookie headers manually
- It makes it impossible to perform GET requests with manually set
headers